+++ holidaysomething/src/main/resources/templates/admin/product/product_search.html
@@ -13,7 +13,6 @@
         <td>
           <select>
             <option value="product_name">상품명</option>
-            <option value="item_name">상품명(관리자)</option>
             <option value="product_code">제품코드(관리자)</option>
           </select>
           <span>
@@ -27,7 +26,7 @@
       </tr>
       <tr>
         <th scope="row">상품분류</th>
-        <td>
+<!--        <td>
           <select name="productCategoryId">
             <option value="">대분류</option>
             <option th:each="big : ${bigCategory}" th:value="${big.id}"
@@ -42,7 +41,7 @@
             <option value="">소분류</option>
           </select>
 
-        </td>
+        </td>-->
       </tr>
       <tr>
         <th scope="row">상품등록일</th>
