+++ holidaysomething/src/main/resources/templates/admin/product/product_search_result.html
@@ -7,12 +7,19 @@
     <h1>조회결과출력페이지!</h1>
     <div>
         <h2>제품검색결과</h2>
-        <table border="1">
-            <th>product name</th>
-            <tr th:each="product : ${productName}"><!--first line-->
-                <td th:text="${product.name}">${product.name}</td>
-            </tr>
-        </table>
+        <!--<table border="1">-->
+            <!--<th>product name</th>-->
+            <!--<tr th:each="product : ${productName}">&lt;!&ndash;first line&ndash;&gt;-->
+                <!--<td th:text="${product.name}">${product.name}</td>-->
+            <!--</tr>-->
+        <!--</table>-->
+
+
+        <ul>
+            <li th:each="product : ${products}">
+                <span th:text="${product.name}"></span>
+            </li>
+        </ul>
 
     </div>
     <br>
@@ -20,8 +27,9 @@
     <!--Pagenation-->
     <div>
         <h2>페이지네이션</h2>
-
-        <span th:text="${presentPage}">${presentPage}</span>
+       <!--&lt;!&ndash; <span th:each="page : ${#numbers.sequence(1, pageCount}">
+        <span th:text="${presentPage}">${presentPage}</span>&ndash;&gt;
+        </span>-->
     </div>
 
 </div>
